requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
backports.zoneinfo==0.2.1
Django==4.2.19
django-redis==5.4.0
django-seed==0.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Faker==35.2.2
pillow==10.4.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
toposort==1.10
typing-extensions==4.12.2
